+++ includes/regular_sorter.hpp
@@ -1,19 +1,3 @@
-// #ifndef TOOLS_HPP
-// #define TOOLS_HPP
-
-// #include <string>
-// #include <vector>
-// #include <utility>
-
-
-// namespace se{
-    
-// void sort_links(std::vector<std::pair<std::string, size_t>>& linksVector, size_t num_res);
-
-// } // namespace se
-
-// #endif
-
 #ifndef REGULAR_SORT_HPP 
 #define REGULAR_SORT_HPP 
 
@@ -35,7 +19,7 @@ public:
     RegularSorter(const RegularSorter&) = default;
     RegularSorter& operator= (const RegularSorter&) = default;
 
-    void sort_links(std::vector<std::pair<std::string, size_t>>& linksVector, size_t num_res)const override;
+    void sort_links(std::vector<std::pair<std::string, int>>& linksVector, size_t num_res)const override;
 };
 
 } //se
